task_1/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
distlib==0.4.0
distro==1.9.0
Django==5.2.7
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
python-dotenv==1.1.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.35.3
Werkzeug==3.1.3
